Rename ITensorVisualization to ITensorVisualizationBase
