+++ .github/workflows/test_itensorunicodeplots.yml
@@ -32,7 +32,7 @@ jobs:
         shell: julia --project=monorepo {0}
         run: |
           using Pkg;
-          pkg"dev ./NDTensors . ./ITensorVisualization ./ITensorUnicodePlots"
+          pkg"dev ./NDTensors . ./ITensorVisualizationBase ./ITensorUnicodePlots"
       - name: Run the tests
         run: >
           julia --project=monorepo --depwarn=yes -e 'using Pkg; Pkg.test("ITensorUnicodePlots")'
