+++ .github/workflows/test_itensorvisualization.yml
@@ -1,4 +1,4 @@
-name: TestITensorVisualization
+name: TestITensorVisualizationBase
 on:
   push:
     branches:
@@ -32,7 +32,7 @@ jobs:
         shell: julia --project=monorepo {0}
         run: |
           using Pkg;
-          pkg"dev ./NDTensors . ./ITensorVisualization"
+          pkg"dev ./NDTensors . ./ITensorVisualizationBase"
       - name: Run the tests
         run: |
-          julia --project=monorepo --depwarn=yes -e 'using Pkg; Pkg.test("ITensorVisualization")'
+          julia --project=monorepo --depwarn=yes -e 'using Pkg; Pkg.test("ITensorVisualizationBase")'
