+++ ITensorMakie/Project.toml
@@ -7,7 +7,7 @@ version = "0.1.0"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 GraphMakie = "1ecd5474-83a3-4783-bb4f-06765db800d2"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-ITensorVisualization = "cd2553d2-8bef-4d93-8a38-c62f17d5ad23"
+ITensorVisualizationBase = "cd2553d2-8bef-4d93-8a38-c62f17d5ad23"
 NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
