+++ ITensorUnicodePlots/Project.toml
@@ -5,7 +5,7 @@ version = "0.1.0"
 
 [deps]
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-ITensorVisualization = "cd2553d2-8bef-4d93-8a38-c62f17d5ad23"
+ITensorVisualizationBase = "cd2553d2-8bef-4d93-8a38-c62f17d5ad23"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
@@ -15,7 +15,7 @@ UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
 [compat]
 Graphs = "1.4.1"
-ITensorVisualization = "0.2.0"
+ITensorVisualizationBase = "0.2.0"
 ITensors = "0.2"
 NetworkLayout = "0.4.3"
 Reexport = "1.2.2"
