+++ ITensorUnicodePlots/test/runtests.jl
@@ -7,7 +7,7 @@ starts_and_ends_with(st, en) = file -> starts_and_ends_with(file, st, en)
 
 test_path = joinpath(@__DIR__)
 test_files = filter(starts_and_ends_with("test_", ".jl"), readdir(test_path))
-@testset "ITensorVisualization.jl" for file in test_files
+@testset "ITensorUnicodePlots.jl" for file in test_files
   file_path = joinpath(test_path, file)
   println("Running test $(file_path)")
   include(file_path)
