+++ ITensorVisualizationBase/Project.toml
@@ -1,4 +1,4 @@
-name = "ITensorVisualization"
+name = "ITensorVisualizationBase"
 uuid = "cd2553d2-8bef-4d93-8a38-c62f17d5ad23"
 authors = ["Matthew Fishman <mfishman@flatironinstitute.org> and contributors"]
 version = "0.2.0"
