+++ ITensorVisualizationBase/examples/ex_2d_tensor_network_layered.jl
@@ -1,5 +1,5 @@
 using ITensors
-using ITensorVisualization
+using ITensorVisualizationBase
 using LayeredLayouts
 using Graphs
