+++ ITensorVisualizationBase/examples/ex_itensor_graph_makie.jl
@@ -1,5 +1,5 @@
 using ITensors
-using ITensorVisualization
+using ITensorVisualizationBase
 using Graphs
 
 g = grid((5,))
