+++ ITensorVisualizationBase/examples/notest_ex_qft_circuit.jl
@@ -1,5 +1,5 @@
 using ITensors
-using ITensorVisualization
+using ITensorVisualizationBase
 using Graphs
 using LayeredLayouts
 using PastaQ: qft
